chore: UserRepositoryTest 제거, UserServiceTest 작성
